Anatomical API

NeuroCore places a strong emphasis on arrays with named dimensions. This allows the use of well established interfaces for arrays across the julia ecosystem while still providing meaningful information that may be specific to the type of data being worked with.

References

NeuroCore.AnatomicalAPI.axial_axis — Method

axial_axis(x, size[; first_pad=nothing, last_pad=nothing, stride=nothing, dilation=nothing])

Returns an AxisIterator along the axial axis.

NeuroCore.AnatomicalAPI.axial_axis — Method

axial_axis(x)

Returns the axis corresponding to the axial dimension.

NeuroCore.AnatomicalAPI.axial_axis_type — Method

axial_axis_type(x)

Returns the key type corresponding to the axial axis.

NeuroCore.AnatomicalAPI.axial_indices — Method

axial_indices(x)

Returns the indices corresponding to the axial axis

NeuroCore.AnatomicalAPI.axial_keys — Method

axial_keys(x)

Returns the keys corresponding to the axial axis

NeuroCore.AnatomicalAPI.axialdim — Method

axialdim(x) -> Int

Returns the dimension corresponding to axial.

NeuroCore.AnatomicalAPI.coronal_axis — Method

coronal_axis(x, size[; first_pad=nothing, last_pad=nothing, stride=nothing, dilation=nothing])

Returns an AxisIterator along the coronal axis.

NeuroCore.AnatomicalAPI.coronal_axis — Method

coronal_axis(x)

Returns the axis corresponding to the coronal dimension.

NeuroCore.AnatomicalAPI.coronal_axis_type — Method

coronal_axis_type(x)

Returns the key type corresponding to the coronal axis.

NeuroCore.AnatomicalAPI.coronal_indices — Method

coronal_indices(x)

Returns the indices corresponding to the coronal axis

NeuroCore.AnatomicalAPI.coronal_keys — Method

coronal_keys(x)

Returns the keys corresponding to the coronal axis

NeuroCore.AnatomicalAPI.coronaldim — Method

coronaldim(x) -> Int

Returns the dimension corresponding to coronal.

NeuroCore.AnatomicalAPI.each_axial — Method

each_axial(x)

Create a generator that iterates over the axial dimensions A, returning views that select all the data from the other dimensions in A.

NeuroCore.AnatomicalAPI.each_coronal — Method

each_coronal(x)

Create a generator that iterates over the coronal dimensions A, returning views that select all the data from the other dimensions in A.

NeuroCore.AnatomicalAPI.each_sagittal — Method

each_sagittal(x)

Create a generator that iterates over the sagittal dimensions A, returning views that select all the data from the other dimensions in A.

NeuroCore.AnatomicalAPI.has_axialdim — Method

has_axialdim(x) -> Bool

Returns true if x has a dimension corresponding to axial.

NeuroCore.AnatomicalAPI.has_coronaldim — Method

has_coronaldim(x) -> Bool

Returns true if x has a dimension corresponding to coronal.

NeuroCore.AnatomicalAPI.has_sagittaldim — Method

has_sagittaldim(x) -> Bool

Returns true if x has a dimension corresponding to sagittal.

NeuroCore.AnatomicalAPI.is_anatomical — Method

is_anatomical(x) -> Bool

Returns true if T represents anatomical data.

NeuroCore.AnatomicalAPI.is_axial — Method

is_axial(x) -> Bool

Returns true if x represent the axial orientation.

NeuroCore.AnatomicalAPI.is_coronal — Method

is_coronal(x) -> Bool

Returns true if x represent the coronal orientation.

NeuroCore.AnatomicalAPI.is_neurologic — Method

is_neurologic(x) -> Bool

Test to see if x is in neurological orientation.

NeuroCore.AnatomicalAPI.is_radiologic — Method

is_radiologic(x) -> Bool

Test to see if x is in radiological orientation.

NeuroCore.AnatomicalAPI.is_sagittal — Method

is_sagittal(x) -> Bool

Returns true if x represent the sagittal orientation.

NeuroCore.AnatomicalAPI.naxial — Method

naxial(x) -> Int

Returns the size along the dimension corresponding to the axial.

NeuroCore.AnatomicalAPI.ncoronal — Method

ncoronal(x) -> Int

Returns the size along the dimension corresponding to the coronal.

NeuroCore.AnatomicalAPI.nsagittal — Method

nsagittal(x) -> Int

Returns the size along the dimension corresponding to the sagittal.

NeuroCore.AnatomicalAPI.sagittal_axis — Method

sagittal_axis(x, size[; first_pad=nothing, last_pad=nothing, stride=nothing, dilation=nothing])

Returns an AxisIterator along the sagittal axis.

NeuroCore.AnatomicalAPI.sagittal_axis — Method

sagittal_axis(x)

Returns the axis corresponding to the sagittal dimension.

NeuroCore.AnatomicalAPI.sagittal_axis_type — Method

sagittal_axis_type(x)

Returns the key type corresponding to the sagittal axis.

NeuroCore.AnatomicalAPI.sagittal_indices — Method

sagittal_indices(x)

Returns the indices corresponding to the sagittal axis

NeuroCore.AnatomicalAPI.sagittal_keys — Method

sagittal_keys(x)

Returns the keys corresponding to the sagittal axis

NeuroCore.AnatomicalAPI.sagittaldim — Method

sagittaldim(x) -> Int

Returns the dimension corresponding to sagittal.

NeuroCore.AnatomicalAPI.EncodingDirection — Type

EncodingDirection

Possible values: ipos, jpos, kpos, ineg,jneg,kneg(the axis of the NIfTI data along which slices were acquired, and the direction in which SliceTiming is defined with respect to).ipos,jpos,kposidentifiers correspond to the first, second and third axis of the data in the NIfTI file.*neg` indicates that the contents of SliceTiming are defined in reverse order - that is, the first entry corresponds to the slice with the largest index, and the final entry corresponds to slice index zero.

NeuroCore.AnatomicalAPI.is_anterior — Method

is_anterior(x) -> Bool

Returns true if x represent the anterior position.

NeuroCore.AnatomicalAPI.is_channel — Method

is_channel(::T) -> Bool

Returns true if T represents a gyrus.

NeuroCore.AnatomicalAPI.is_cortical — Method

is_cortical(::T) -> Bool

Returns true if T represents a cortical region.

NeuroCore.AnatomicalAPI.is_csf — Method

is_csf(::T) -> Bool

Returns true if T represents a region of corticospinal fluid (CSF).

NeuroCore.AnatomicalAPI.is_electrophysiology — Method

is_electrophysiology(::T) -> Bool

Returns true if T represents electrophysiology data.

NeuroCore.AnatomicalAPI.is_gyrus — Method

is_gyrus(::T) -> Bool

Returns true if T represents a gyrus.

NeuroCore.AnatomicalAPI.is_inferior — Method

is_inferior(x) -> Bool

Returns true if x represent the inferior position.

NeuroCore.AnatomicalAPI.is_left — Method

is_left(x) -> Bool

Returns true if x represent the left position.

NeuroCore.AnatomicalAPI.is_posterior — Method

is_posterior(x) -> Bool

Returns true if x represent the posterior position.

NeuroCore.AnatomicalAPI.is_right — Method

is_right(x) -> Bool

Returns true if x represent the right position.

NeuroCore.AnatomicalAPI.is_sulcus — Method

is_sulcus(::T) -> Bool

Returns true if T represents a gyrus.

NeuroCore.AnatomicalAPI.is_superior — Method

is_superior(x) -> Bool

Returns true if x represent the superior position.

NeuroCore.AnatomicalAPI.is_white_matter — Method

is_white_matter(::T) -> Bool

Returns true if T represents white matter.

NeuroCore.AnatomicalAPI.select_axial — Method

select_axial(x, i)

Return a view of all the data of x where the index for the axial dimension equals i.

NeuroCore.AnatomicalAPI.select_coronal — Method

select_coronal(x, i)

Return a view of all the data of x where the index for the coronal dimension equals i.

NeuroCore.AnatomicalAPI.select_sagittal — Method

select_sagittal(x, i)

Return a view of all the data of x where the index for the sagittal dimension equals i.