Orientation and Dimension Names

NeuroCore places a strong emphasis on arrays that with named dimensions. This allows the use of well established interfaces for arrays across the julia ecosystem while still providing meaningful information that may be specific to the type of data being worked with.

NeuroCore.sagittaldim — Function

sagittaldim(x) -> Int

Return the dimension of the array used for sagittal time.

NeuroCore.coronaldim — Function

coronaldim(x) -> Int

Return the dimension of the array used for coronal time.

NeuroCore.axialdim — Function

axialdim(x) -> Int

Return the dimension of the array used for axial time.

NeuroCore.indices_sagittal — Function

indices_sagittal(x)

Return the indices of the sagittal dimension.

NeuroCore.indices_axial — Function

indices_axial(x)

Return the indices of the axial dimension.

NeuroCore.indices_coronal — Function

indices_coronal(x)

Return the indices of the coronal dimension.

Missing docstring.

Missing docstring for indices_unit. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Onset. Check Documenter's build log for details.

Missing docstring.

Missing docstring for StopTime. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Duration. Check Documenter's build log for details.

Missing docstring.

Missing docstring for TimeContinuity. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SamplingRate. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SpatialOffset. Check Documenter's build log for details.

Missing docstring.

Missing docstring for spatial_units. Check Documenter's build log for details.

NeuroCore.is_radiologic — Function

is_radiologic(x) -> Bool

Test to see if x is in radiological orientation.

NeuroCore.is_neurologic — Function

is_neurologic(x) -> Bool

Test to see if x is in neurological orientation.