Imaging Metadata

Sequence Metadata

Missing docstring.

Missing docstring for NeuroCore.SequenceMetadata. Check Documenter's build log for details.

Properties

Missing docstring.

Missing docstring for NeuroCore.nonlinear_gradient_correction. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NeuroCore.pulse_sequence. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NeuroCore.pulse_sequence_details. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NeuroCore.scanning_sequence. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NeuroCore.sequence_name. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NeuroCore.sequence_variant. Check Documenter's build log for details.

Encoding Direction

NeuroCore.NeuroMetadata.EncodingDirectionMetadata — Type

EncodingDirectionMetadata

Metadata structure for general MRI sequence information.

Supported Properties

freqdim: Which spatial dimension (1, 2, or 3) corresponds to phase acquisition.

phasedim: Which spatial dimension (1, 2, or 3) corresponds to phase acquisition.

slicedim: Which dimension slices where acquired at throughout MRI acquisition.

slice_start: Which slice corresponds to the first slice acquired during MRI acquisition (i.e. not padded slices).

slice_end: Which slice corresponds to the last slice acquired during MRI acquisition (i.e. not padded slices).

slice_duration: Time to acquire one slice

slice_encoding_direction: Values ending in "neg" indicate that the contents of slicetiming are defined in reverse order (the first entry corresponds to the slice with the largest index, and the final entry corresponds to slice index zero. When present, the axis defined by slice*encodingneeds to be consistent with theslicedim field in the NIfTI header. When absent, the entries in slice_timing must be in the order of increasing slice index as defined by the NIfTI header.

phase_encoding_direction: The phase encoding direction is defined as the direction along which phase was modulated which may result in visible distortions. Note that this is not the same as the DICOM term "inplanephaseencodingdirection" which can have "ROW" or "COL" values. This parameter is REQUIRED if corresponding fieldmap data is present or when using multiple runs with different phase encoding directions (which can be later used for field inhomogeneity correction).

Examples

julia> using NeuroCore

julia> using NeuroCore.NeuroMetadata

julia> m = EncodingDirectionMetadata(1, 2, 3, 4, 5, 6)
EncodingDirectionMetadata(1, 2, 3, 4, 5, 6.0)

julia> m.slice_encoding_direction
kpos::EncodingDirection = 3

julia> m.slice_encoding_direction == slice_encoding_direction(m)
true

julia> m.phase_encoding_direction
jpos::EncodingDirection = 2

julia> m.phase_encoding_direction == phase_encoding_direction(m)
true

julia> m.freqdim
1

julia> m.freqdim == freqdim(m)
true

julia> m.phasedim
2

julia> m.phasedim == phasedim(m)
true

julia> m.slicedim
3

julia> m.slicedim == slicedim(m)
true

julia> m.slice_start
4

julia> m.slice_start == slice_start(m)
true

julia> m.slice_end
5

julia> m.slice_end == slice_end(m)
true

julia> m.slice_duration
6.0

julia> m.slice_duration == slice_duration(m)
true

Properties

Missing docstring.

Missing docstring for NeuroCore.EncodingDirection. Check Documenter's build log for details.

NeuroCore.NeuroMetadata.phase_encoding_direction — Function

phase_encoding_direction(x) -> EncodingDirection

The phase encoding direction is defined as the direction along which phase was modulated which may result in visible distortions. Note that this is not the same as the DICOM term "inplanephaseencodingdirection" which can have "ROW" or "COL" values. This parameter is REQUIRED if corresponding fieldmap data is present or when using multiple runs with different phase encoding directions (which can be later used for field inhomogeneity correction).

NeuroCore.NeuroMetadata.slice_encoding_direction — Function

slice_encoding_direction(x) -> EncodingDirection

Values ending in "*neg" indicate that the contents of slicetiming are defined in reverse order (the first entry corresponds to the slice with the largest index, and the final entry corresponds to slice index zero. When present, the axis defined by `sliceencodingneeds to be consistent with theslicedim` field in the NIfTI header. When absent, the entries in slice_timing must be in the order of increasing slice index as defined by the NIfTI header.

NeuroCore.NeuroMetadata.freqdim — Function

freqdim(x)
freqdim!(x, val)

Which spatial dimension (1, 2, or 3) corresponds to phase acquisition.

NeuroCore.NeuroMetadata.phasedim — Function

phasedim(x)
phasedim!(x, val)

Which spatial dimension (1, 2, or 3) corresponds to phase acquisition.

NeuroCore.NeuroMetadata.slicedim — Function

slicedim(x)
slicedim!(x, val)

Which dimension slices where acquired at throughout MRI acquisition.

NeuroCore.NeuroMetadata.slice_start — Function

slice_start(x)
slice_start!(x, val)

Which slice corresponds to the first slice acquired during MRI acquisition (i.e. not padded slices).

NeuroCore.NeuroMetadata.slice_end — Function

slice_end(x)
slice_end!(x, val)

Which slice corresponds to the last slice acquired during MRI acquisition (i.e. not padded slices).

NeuroCore.NeuroMetadata.slice_duration — Function

slice_duration(x)
slice_duration!(x, val)

Time to acquire one slice

Spatial Encoding

Properties

Missing docstring.

Missing docstring for NeuroCore.nshots. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NeuroCore.effective_echo_spacing. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NeuroCore.parallel_acquisition_technique. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NeuroCore.parallel_reduction_factor_in_plane. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NeuroCore.partial_fourier. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NeuroCore.partial_fourier_direction. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NeuroCore.total_readout_time. Check Documenter's build log for details.

Magentization Transfer

Missing docstring.

Missing docstring for NeuroCore.MagnetizationTransferMetadata. Check Documenter's build log for details.

Properties

Missing docstring.

Missing docstring for NeuroCore.mt_state. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NeuroCore.mt_offset_frequency. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NeuroCore.mt_pulse_bandwidth. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NeuroCore.mt_npulses. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NeuroCore.mt_pulse_shape. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NeuroCore.mt_pulse_duration. Check Documenter's build log for details.

Spoiling

Missing docstring.

Missing docstring for NeuroCore.SpoilingMetadata. Check Documenter's build log for details.

Properties

Missing docstring.

Missing docstring for NeuroCore.spoiling_state. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NeuroCore.spoiling_type. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NeuroCore.spoiling_gradient_moment. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NeuroCore.spoiling_gradient_duration. Check Documenter's build log for details.

Time

Missing docstring.

Missing docstring for NeuroCore.echo_time. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NeuroCore.inversion_time. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NeuroCore.slice_timing. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NeuroCore.dwell_time. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NeuroCore.delay_time. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NeuroCore.acquisition_duration. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NeuroCore.volume_timing. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NeuroCore.repetition_time. Check Documenter's build log for details.