NeuroCore Types

The NeuroArray

Warning

Although the basic syntax for interacting with a NeuroArray is not expected to change, there are numerous efforts underway to simplify and optimizae the arrays with named dimensions and special indices. Therefore, some of the following documentation may change in the future.

Julia benefits enormously from having a common interface to customizable arrays and using the AbstractArray interface instantly allows us access to cutting edge algorithms that would otherwise take months to years to make compatible with neuroscience specific formats. NeuroCore takes ownership over arrays from existing packages (e.g., ImageMeta from ImageMetadata) by including a custom metadata field. In so doing we inherit well maintained array structures and can create unique methods when necessary.

The Orientation and Dimension Names section covers unique methods made available through this package for dimensions. The remaining documentation is mostly dedicated to accessing and composing various forms of metadata.

NeuroCore.NeuroArray — Type

NeuroArray(array; axes...[, metadata])

Examples

julia> using NeuroCore

julia> x = NeuroArray(rand(2,3,4,10);
               left = 1:2,
               anterior = 1:3,
               superior=1:4,
               time = range(1, stop=10, length=10));

julia> dimnames(x)
(:left, :anterior, :superior, :time)

julia> is_radiologic(x)
true

julia> is_neurologic(x)
false

julia> dim(x, :left) == sagittaldim(x) == 1
true

julia> coronaldim(x)
2

julia> axialdim(x)
3

julia> indices_sagittal(x)
1:2

julia> indices_coronal(x)
1:3

julia> indices_axial(x)
1:4

julia> sampling_rate(x)
1.0

julia> duration(x)
9.0

julia> stop_time(x)
10.0

Coordinates

Sets of coordinates may be created using a vector of tuples through the NeuroCoordinates type. This offers the same functionality as a NeuroArray for storing unique metadata.

NeuroCore.NeuroCoordinates — Type

NeuroCoordinates

Examples

julia> using NeuroCore

julia> x = NeuroCoordinates([(1,1,1), (1,1,2)], [:roi1, :roi2]);