Orientation and Dimension Names

NeuroCore places a strong emphasis on arrays with named dimensions. This allows the use of well established interfaces for arrays across the julia ecosystem while still providing meaningful information that may be specific to the type of data being worked with.

NeuroCore.sagittaldim — Function

sagittaldim(x) -> Int

Return the dimension of the array used for sagittal time.

NeuroCore.coronaldim — Function

coronaldim(x) -> Int

Return the dimension of the array used for coronal time.

NeuroCore.axialdim — Function

axialdim(x) -> Int

Return the dimension of the array used for axial time.

NeuroCore.indices_sagittal — Function

indices_sagittal(x)

Return the indices of the sagittal dimension.

NeuroCore.indices_axial — Function

indices_axial(x)

Return the indices of the axial dimension.

NeuroCore.indices_coronal — Function

indices_coronal(x)

Return the indices of the coronal dimension.

NeuroCore.onset — Function

onset(x)
onset!(x, val)

First time point along the time axis.

NeuroCore.stop_time — Function

Last time point along the time axis.

NeuroCore.duration — Function

Duration of the event along the time axis.

NeuroCore.time_continuity — Function

time_continuity(x)
time_continuity!(x, val)

Defines whether the recording is "continuous", "discontinuous" or "epoched"; this latter limited to time windows about events of interest (e.g., stimulus presentations, subject responses etc.)

NeuroCore.sampling_rate — Function

sampling_rate(x)
sampling_rate!(x, val)

Number of samples per second.

NeuroCore.spatial_offset — Function

spatial_offset(x)
spatial_offset!(x, val)

The offset of each dimension (i.e., where each spatial axis starts).

NeuroCore.spatial_units — Function

spatial_units(x)

Returns the units (i.e. Unitful.unit) that each spatial axis is measured in. If not available nothing is returned for each spatial axis.

NeuroCore.is_radiologic — Function

is_radiologic(x) -> Bool

Test to see if x is in radiological orientation.

NeuroCore.is_neurologic — Function

is_neurologic(x) -> Bool

Test to see if x is in neurological orientation.