Imaging Metadata

Sequence Metadata

Properties

NeuroCore.NonlinearGradientCorrection — Type

Returns Bool stating if the image saved has been corrected for gradient nonlinearities by the scanner sequence. Default is false.

NeuroCore.PulseSequence — Type

General description of the pulse sequence used for the scan (i.e. MPRAGE, Gradient Echo EPI, Spin Echo EPI, Multiband gradient echo EPI).

NeuroCore.PulseSequenceDetails — Type

Information beyond pulse sequence type that identifies the specific pulse sequence used (i.e. "Standard Siemens Sequence distributed with the VB17 software," "Siemens WIP ### version #.##," or "Sequence written by X using a version compiled on MM/DD/YYYY").

NeuroCore.PulseSequenceType — Type

A general description of the pulse sequence used for the scan (i.e. MPRAGE, Gradient Echo EPI, Spin Echo EPI, Multiband gradient echo EPI).

NeuroCore.ScanningSequence — Type

Description of the type of sequence data acquired.

NeuroCore.SequenceName — Type

Manufacturer’s designation of the sequence name.

NeuroCore.SequenceVarient — Type

Variant of the scanning_sequence property.

Metadata

NeuroCore.SequenceMetadata — Type

SequenceMetadata

Metadata structure for general MRI sequence information.

Encoding Direction

NeuroCore.EncodingDirections.EncodingDirection — Type

EncodingDirection

Possible values: i, j, k, ineg,jneg,kneg(the axis of the NIfTI data along which slices were acquired, and the direction in which SliceTiming is defined with respect to).i,j,kidentifiers correspond to the first, second and third axis of the data in the NIfTI file.*neg` indicates that the contents of SliceTiming are defined in reverse order - that is, the first entry corresponds to the slice with the largest index, and the final entry corresponds to slice index zero.

NeuroCore.EncodingDirections.EncodingDirectionMetadata — Type

EncodingDirectionMetadata

Metadata structure for general MRI sequence information.

Properties

freqdim: Which spatial dimension (1, 2, or 3) corresponds to phase acquisition.

phasedim: Which spatial dimension (1, 2, or 3) corresponds to phase acquisition.

slicedim: Which dimension slices where acquired at throughout MRI acquisition.

slice_start: Which slice corresponds to the first slice acquired during MRI acquisition (i.e. not padded slices).

slice_end: Which slice corresponds to the last slice acquired during MRI acquisition (i.e. not padded slices).

slice_duration: Time to acquire one slice

slice_encoding_direction: Possible values:

phase_encoding_direction: The phase encoding direction is defined as the direction along which phase was modulated which may result in visible distortions. Note that this is not the same as the DICOM term

Examples

julia> using NeuroCore.EncodingDirections

julia> m = EncodingDirectionMetadata(1, 2, 3, 4, 5, 6)
EncodingDirectionMetadata(1, 2, 3, 4, 5, 6.0)

julia> m.slice_encoding_direction
kpos::EncodingDirection = 3

julia> m.phase_encoding_direction
jpos::EncodingDirection = 2

julia> m.freqdim
1

julia> m.phasedim
2

julia> m.slicedim
3

julia> m.slice_start
4

julia> m.slice_end
5

julia> m.slice_duration
6.0

NeuroCore.EncodingDirections.FrequencyDimension — Type

Which spatial dimension (1, 2, or 3) corresponds to phase acquisition.

NeuroCore.EncodingDirections.PhaseDimension — Type

Which spatial dimension (1, 2, or 3) corresponds to phase acquisition.

NeuroCore.EncodingDirections.PhaseEncodingDirection — Type

The phase encoding direction is defined as the direction along which phase was modulated which may result in visible distortions. Note that this is not the same as the DICOM term in_plane_phase_encoding_direction which can have ROW or COL values. This parameter is REQUIRED if corresponding fieldmap data is present or when using multiple runs with different phase encoding directions (which can be later used for field inhomogeneity correction).

NeuroCore.EncodingDirections.SliceDimension — Type

Which dimension slices where acquired at throughout MRI acquisition.

NeuroCore.EncodingDirections.SliceDuration — Type

Time to acquire one slice

NeuroCore.EncodingDirections.SliceEncodingDirection — Type

Possible values: i, j, k, ineg,jneg,kneg(the axis of the NIfTI data along which slices were acquired, and the direction in which slice_timing is defined with respect to).i,j,kidentifiers correspond to the first, second and third axis of the data in the NIfTI file.*negindicates that the contents of slice_timing are defined in reverse order - that is, the first entry corresponds to the slice with the largest index, and the final entry corresponds to slice index zero. When present, the axis defined bysliceencoding` needs to be consistent with the ‘slicedim’ field in the NIfTI header. When absent, the entries in slicetiming must be in the order of increasing slice index as defined by the NIfTI header.

NeuroCore.EncodingDirections.SliceEnd — Type

Which slice corresponds to the last slice acquired during MRI acquisition (i.e. not padded slices).

NeuroCore.EncodingDirections.SliceStart — Type

Which slice corresponds to the first slice acquired during MRI acquisition (i.e. not padded slices).

Spatial Encoding

Properties

NeuroCore.NumberShots — Type

The number of RF excitations needed to reconstruct a slice or volume. Please mind that this is not the same as Echo Train Length which denotes the number of lines of k-space collected after an excitation.

NeuroCore.EffectiveEchoSpacing — Type

The "effective" sampling interval, specified in seconds, between lines in the phase-encoding direction, defined based on the size of the reconstructed image in the phase direction. It is frequently, but incorrectly, referred to as "dwell time" (see dwell_time parameter below for actual dwell time). It is required for unwarping distortions using field maps. Note that beyond just in-plane acceleration, a variety of other manipulations to the phase encoding need to be accounted for properly, including partial fourier, phase oversampling, phase resolution, phase field-of-view and interpolation.<sup>2</sup> This parameter is REQUIRED if corresponding fieldmap data is present.

<sup>2</sup>Conveniently, for Siemens’ data, this value is easily obtained as 1/[BWPPPE * ReconMatrixPE], where BWPPPE is the "BandwidthPerPixelPhaseEncode in DICOM tag (0019,1028) and ReconMatrixPE is the size of the actual reconstructed data in the phase direction (which is NOT reflected in a single DICOM tag for all possible aforementioned scan manipulations). See here and here

NeuroCore.ParallelAcquisitionTechnique — Type

The type of parallel imaging used (e.g. GRAPPA, SENSE).

NeuroCore.ParallelReductionFactor — Type

The parallel imaging (e.g, GRAPPA) factor. Use the denominator of the fraction of k-space encoded for each slice. For example, 2 means half of k-space is encoded.

NeuroCore.PartialFourier — Type

The fraction of partial Fourier information collected.

NeuroCore.PartialFourierDirection — Type

The direction where only partial Fourier information was collected.

NeuroCore.TotalReadoutTime — Type

The total readout time. This is actually the "effective" total readout time , defined as the readout duration, specified in seconds, that would have generated data with the given level of distortion. It is NOT the actual, physical duration of the readout train. If effective_echo_spacing has been properly computed, it is just effective_echo_spacing * (ReconMatrixPE - 1).<sup>3</sup> .

This parameter is

REQUIRED if corresponding "field/distortion" maps acquired with opposing phase encoding directions are present.

<sup>3</sup>We use the "FSL definition", i.e, the time between the center of the first "effective" echo and the center of the last "effective" echo.

Magentization Transfer

Properties

NeuroCore.MTState — Type

Specifies whether the magnetization transfer pulse is applied. This parameter is REQUIRED by all the anatomical images grouped by MTR, MTS and MPM suffixes.

NeuroCore.MTOffsetFrequency — Type

The frequency offset of the magnetization transfer pulse with respect to the central H1 Larmor frequency in Hertz (Hz).

NeuroCore.MTPulseBandwidth — Type

The excitation bandwidth of the magnetization transfer pulse in Hertz (Hz).

NeuroCore.MTNumberOfPulses — Type

Number of magnetization transfer RF pulses applied before the readout.

NeuroCore.MTPulseShape — Type

Shape of the magnetization transfer RF pulse waveform. Accepted values:

HARD

GAUSSIAN

GAUSSHANN (gaussian pulse with Hanning window)

SINC

SINCHANN (sinc pulse with Hanning window)

SINCGAUSS (sinc pulse with Gaussian window)

FERMI

NeuroCore.MTPulseDuration — Type

Duration of the magnetization transfer RF pulse in seconds.

Metadata

NeuroCore.MagnetizationTransferMetadata — Type

MagnetizationTransferMetadata

Metadata structure for information concerning MRI magnetization transfer pulse.

Spoiling

Properties

NeuroCore.SpoilingState — Type

Specifies whether the pulse sequence uses any type of spoiling stratey to suppress transverse magnetization remaining after the readout.

NeuroCore.SpoilingType — Type

Specifies which spoiling method(s) are used by a spoiled sequence. Accepted values: RF, GRADIENT or COMBINED.

NeuroCore.SpoilingGradientMoment — Type

Zeroth moment of the spoiler gradient lobe in militesla times second per meter (mT.s/m).

NeuroCore.SpoilingGradientDuration — Type

The duration of the spoiler gradient lobe in seconds. The duration of a trapezoidal lobe is defined as the summation of ramp-up and plateu times.

Metadata

NeuroCore.SpoilingMetadata — Type

SpoilingMetadata

Metadata structure for information concerning MRI sequence spoiling.

Time

NeuroCore.EchoTime — Type

The echo time (TE) for the acquisition. This parameter is REQUIRED if corresponding fieldmap data is present or the data comes from a multi echo sequence.

NeuroCore.InversionTime — Type

Returns the inversion time (TI) for the acquisition, specified in seconds. Inversion time is the time after the middle of inverting RF pulse to middle of excitation pulse to detect the amount of longitudinal magnetization.

NeuroCore.SliceTiming — Type

The time at which each slice was acquired within each volume (frame) of the acquisition. Slice timing is not slice order – rather, it is a list of times (in JSON format) containing the time (in seconds) of each slice acquisition in relation to the beginning of volume acquisition. The list goes through the slices along the slice axis in the slice encoding dimension (see below). Note that to ensure the proper interpretation of the slice_timing field, it is important to check if the OPTIONAL slice_encoding_direction exists. In particular, if slice_encoding_direction is negative, the entries in slice_timing are defined in reverse order with respect to the slice axis (i.e., the final entry in the slice_timing list is the time of acquisition of slice 0). This parameter is REQUIRED for sparse sequences that do not have the delay_time field set. In addition without this parameter slice time correction will not be possible.

NeuroCore.DwellTime — Type

Actual dwell time (in seconds) of the receiver per point in the readout direction, including any oversampling. For Siemens, this corresponds to DICOM field (0019,1018) (in ns). This value is necessary for the optional readout distortion correction of anatomicals in the HCP Pipelines. It also usefully provides a handle on the readout bandwidth, which isn’t captured in the other metadata tags. Not to be confused with effective_echo_spacing, and the frequent mislabeling of echo spacing (which is spacing in the phase encoding direction) as "dwell time" (which is spacing in the readout direction).

NeuroCore.DelayTime — Type

Returns the user specified time (in seconds) to delay the acquisition of data for the following volume. If the field is not present it is assumed to be set to zero. Corresponds to Siemens CSA header field ldelaytimeInTR. This field is REQUIRED for sparse sequences using the repetitiontime field that do not have the slicetiming field set to allowed for accurate calculation of "acquisition time". This field is mutually exclusive with `volumetiming`.

NeuroCore.AcquisitionDuration — Type

Duration (in seconds) of volume acquisition. This field is REQUIRED for sequences that are described with the volume_timingfield and that do not have the slice_timing field set to allowed for accurate calculation of "acquisition time". This field is mutually exclusive with repetition_time.

NeuroCore.VolumeTiming — Type

Returns the time at which each volume was acquired during the acquisition. It is described using a list of times (in JSON format) referring to the onset of each volume in the BOLD series. The list must have the same length as the BOLD series, and the values must be non-negative and monotonically increasing. This field is mutually exclusive with repetitiontime and delaytime. If defined, this requires acquisition time (TA) be defined via either slicetiming or acquisitionduration be defined.

NeuroCore.RepetitionTime — Type

Returns the time in seconds between the beginning of an acquisition of one volume and the beginning of acquisition of the volume following it (TR). Please note that this definition includes time between scans (when no data has been acquired) in case of sparse acquisition schemes. This value needs to be consistent with the pixdim[4] field (after accounting for units stored in xyztunits field) in the NIfTI header. This field is mutually exclusive with `volumetiming` and is derived from DICOM Tag 0018, 0080 and converted to seconds.