Imaging Metadata

Sequence Metadata

Properties

NeuroCore.NonlinearGradientCorrection — Type

Returns Bool stating if the image saved has been corrected for gradient nonlinearities by the scanner sequence. Default is false.

NeuroCore.PulseSequence — Type

General description of the pulse sequence used for the scan (i.e. MPRAGE, Gradient Echo EPI, Spin Echo EPI, Multiband gradient echo EPI).

NeuroCore.PulseSequenceDetails — Type

Information beyond pulse sequence type that identifies the specific pulse sequence used (i.e. "Standard Siemens Sequence distributed with the VB17 software," "Siemens WIP ### version #.##," or "Sequence written by X using a version compiled on MM/DD/YYYY").

NeuroCore.PulseSequenceType — Type

A general description of the pulse sequence used for the scan (i.e. MPRAGE, Gradient Echo EPI, Spin Echo EPI, Multiband gradient echo EPI).

NeuroCore.ScanningSequence — Type

Description of the type of sequence data acquired.

NeuroCore.SequenceName — Type

Manufacturer’s designation of the sequence name.

NeuroCore.SequenceVarient — Type

Variant of the scanning_sequence property.

Metadata

NeuroCore.SequenceMetadata — Type

SequenceMetadata

Metadata structure for general MRI sequence information.

Encoding Direction

Properties

NeuroCore.EncodingDirection — Type

EncodingDirection

Possible values: i, j, k, ineg,jneg,kneg(the axis of the NIfTI data along which slices were acquired, and the direction in which SliceTiming is defined with respect to).i,j,kidentifiers correspond to the first, second and third axis of the data in the NIfTI file.*neg` indicates that the contents of SliceTiming are defined in reverse order - that is, the first entry corresponds to the slice with the largest index, and the final entry corresponds to slice index zero.

NeuroCore.PhaseEncodingDirection — Type

The phase encoding direction is defined as the direction along which phase is was modulated which may result in visible distortions. Note that this is not the same as the DICOM term in_plane_phase_encoding_direction which can have ROW or COL values. This parameter is REQUIRED if corresponding fieldmap data is present or when using multiple runs with different phase encoding directions (which can be later used for field inhomogeneity correction).

NeuroCore.SliceEncodingDirection — Type

Possible values: i, j, k, ineg,jneg,kneg(the axis of the NIfTI data along which slices were acquired, and the direction in which slice_timing is defined with respect to).i,j,kidentifiers correspond to the first, second and third axis of the data in the NIfTI file.*negindicates that the contents of slice_timing are defined in reverse order - that is, the first entry corresponds to the slice with the largest index, and the final entry corresponds to slice index zero. When present, the axis defined bysliceencoding` needs to be consistent with the ‘slicedim’ field in the NIfTI header. When absent, the entries in slicetiming must be in the order of increasing slice index as defined by the NIfTI header.

NeuroCore.FrequencyDimension — Type

Which spatial dimension (1, 2, or 3) corresponds to phase acquisition.

NeuroCore.PhaseDimension — Type

Which spatial dimension (1, 2, or 3) corresponds to phase acquisition.

NeuroCore.SliceStart — Type

Which slice corresponds to the first slice acquired during MRI acquisition (i.e. not padded slices).

NeuroCore.SliceEnd — Type

Which slice corresponds to the last slice acquired during MRI acquisition (i.e. not padded slices).

NeuroCore.SliceDim — Type

Which dimension slices where acquired at throughout MRI acquisition.

NeuroCore.SliceDuration — Type

Time to acquire one slice

Metadata

NeuroCore.EncodingDirectionMetadata — Type

EncodingDirectionMetadata

Metadata structure for general MRI sequence information.

Magentization Transfer

Properties

NeuroCore.MTState — Type

Specifies whether the magnetization transfer pulse is applied. This parameter is REQUIRED by all the anatomical images grouped by MTR, MTS and MPM suffixes.

NeuroCore.MTOffsetFrequency — Type

The frequency offset of the magnetization transfer pulse with respect to the central H1 Larmor frequency in Hertz (Hz).

NeuroCore.MTPulseBandwidth — Type

The excitation bandwidth of the magnetization transfer pulse in Hertz (Hz).

NeuroCore.MTNumberOfPulses — Type

Number of magnetization transfer RF pulses applied before the readout.

NeuroCore.MTPulseShape — Type

Shape of the magnetization transfer RF pulse waveform. Accepted values:

HARD

GAUSSIAN

GAUSSHANN (gaussian pulse with Hanning window)

SINC

SINCHANN (sinc pulse with Hanning window)

SINCGAUSS (sinc pulse with Gaussian window)

FERMI

NeuroCore.MTPulseDuration — Type

Duration of the magnetization transfer RF pulse in seconds.

Metadata

NeuroCore.MagnetizationTransferMetadata — Type

MagnetizationTransferMetadata

Metadata structure for information concerning MRI magnetization transfer pulse.