Imaging Metadata

Sequence Metadata

NeuroCore.SequenceMetadata — Type

SequenceMetadata

Metadata structure for general MRI sequence information.

Supported Properties

nonlinear_gradient_correction: Returns Bool stating if the image saved has been corrected for gradient nonlinearities by the scanner sequence. Default is false.

pulse_sequence: General description of the pulse sequence used for the scan (i.e. MPRAGE, Gradient Echo EPI, Spin Echo EPI, Multiband gradient echo EPI).

pulse_sequence_details: Information beyond pulse sequence type that identifies the specific pulse sequence used (i.e. "Standard Siemens Sequence distributed with the VB17 software," "Siemens WIP ### version #.##," or "Sequence written by X using a version compiled on MM/DD/YYYY").

pulse_sequence: General description of the pulse sequence used for the scan (i.e. MPRAGE, Gradient Echo EPI, Spin Echo EPI, Multiband gradient echo EPI).

scanning_sequence: Description of the type of sequence data acquired.

sequence_name: Manufacturer’s designation of the sequence name.

sequence_varient: Variant of the scanning_sequence property.

Properties

NeuroCore.nonlinear_gradient_correction — Function

nonlinear_gradient_correction(x)
nonlinear_gradient_correction!(x, val)

Returns Bool stating if the image saved has been corrected for gradient nonlinearities by the scanner sequence. Default is false.

NeuroCore.pulse_sequence — Function

pulse_sequence(x)
pulse_sequence!(x, val)

General description of the pulse sequence used for the scan (i.e. MPRAGE, Gradient Echo EPI, Spin Echo EPI, Multiband gradient echo EPI).

NeuroCore.pulse_sequence_details — Function

pulse_sequence_details(x)
pulse_sequence_details!(x, val)

Information beyond pulse sequence type that identifies the specific pulse sequence used (i.e. "Standard Siemens Sequence distributed with the VB17 software," "Siemens WIP ### version #.##," or "Sequence written by X using a version compiled on MM/DD/YYYY").

NeuroCore.scanning_sequence — Function

scanning_sequence(x)
scanning_sequence!(x, val)

Description of the type of sequence data acquired.

NeuroCore.sequence_name — Function

sequence_name(x)
sequence_name!(x, val)

Manufacturer’s designation of the sequence name.

NeuroCore.sequence_varient — Function

sequence_varient(x)
sequence_varient!(x, val)

Variant of the scanning_sequence property.

Encoding Direction

NeuroCore.EncodingDirectionMetadata — Type

EncodingDirectionMetadata

Metadata structure for general MRI sequence information.

Supported Properties

freqdim: Which spatial dimension (1, 2, or 3) corresponds to phase acquisition.

phasedim: Which spatial dimension (1, 2, or 3) corresponds to phase acquisition.

slicedim: Which dimension slices where acquired at throughout MRI acquisition.

slice_start: Which slice corresponds to the first slice acquired during MRI acquisition (i.e. not padded slices).

slice_end: Which slice corresponds to the last slice acquired during MRI acquisition (i.e. not padded slices).

slice_duration: Time to acquire one slice

slice_encoding_direction: Values ending in "neg" indicate that the contents of slicetiming are defined in reverse order (the first entry corresponds to the slice with the largest index, and the final entry corresponds to slice index zero. When present, the axis defined by slice*encodingneeds to be consistent with theslicedim field in the NIfTI header. When absent, the entries in slice_timing must be in the order of increasing slice index as defined by the NIfTI header.

phase_encoding_direction: The phase encoding direction is defined as the direction along which phase was modulated which may result in visible distortions. Note that this is not the same as the DICOM term "inplanephaseencodingdirection" which can have "ROW" or "COL" values. This parameter is REQUIRED if corresponding fieldmap data is present or when using multiple runs with different phase encoding directions (which can be later used for field inhomogeneity correction).

Examples

julia> using NeuroCore

julia> m = EncodingDirectionMetadata(1, 2, 3, 4, 5, 6)
EncodingDirectionMetadata(1, 2, 3, 4, 5, 6.0)

julia> m.slice_encoding_direction
kpos::EncodingDirection = 3

julia> m.slice_encoding_direction == slice_encoding_direction(m)
true

julia> m.phase_encoding_direction
jpos::EncodingDirection = 2

julia> m.phase_encoding_direction == phase_encoding_direction(m)
true

julia> m.freqdim
1

julia> m.freqdim == freqdim(m)
true

julia> m.phasedim
2

julia> m.phasedim == phasedim(m)
true

julia> m.slicedim
3

julia> m.slicedim == slicedim(m)
true

julia> m.slice_start
4

julia> m.slice_start == slice_start(m)
true

julia> m.slice_end
5

julia> m.slice_end == slice_end(m)
true

julia> m.slice_duration
6.0

julia> m.slice_duration == slice_duration(m)
true

Properties

NeuroCore.EncodingDirection — Type

EncodingDirection

Possible values: ipos, jpos, kpos, ineg,jneg,kneg(the axis of the NIfTI data along which slices were acquired, and the direction in which SliceTiming is defined with respect to).ipos,jpos,kposidentifiers correspond to the first, second and third axis of the data in the NIfTI file.*neg` indicates that the contents of SliceTiming are defined in reverse order - that is, the first entry corresponds to the slice with the largest index, and the final entry corresponds to slice index zero.

NeuroCore.phase_encoding_direction — Function

phase_encoding_direction(x) -> EncodingDirection

The phase encoding direction is defined as the direction along which phase was modulated which may result in visible distortions. Note that this is not the same as the DICOM term "inplanephaseencodingdirection" which can have "ROW" or "COL" values. This parameter is REQUIRED if corresponding fieldmap data is present or when using multiple runs with different phase encoding directions (which can be later used for field inhomogeneity correction).

NeuroCore.slice_encoding_direction — Function

slice_encoding_direction(x) -> EncodingDirection

Values ending in "*neg" indicate that the contents of slicetiming are defined in reverse order (the first entry corresponds to the slice with the largest index, and the final entry corresponds to slice index zero. When present, the axis defined by `sliceencodingneeds to be consistent with theslicedim` field in the NIfTI header. When absent, the entries in slice_timing must be in the order of increasing slice index as defined by the NIfTI header.

NeuroCore.freqdim — Function

freqdim(x)
freqdim!(x, val)

Which spatial dimension (1, 2, or 3) corresponds to phase acquisition.

NeuroCore.phasedim — Function

phasedim(x)
phasedim!(x, val)

Which spatial dimension (1, 2, or 3) corresponds to phase acquisition.

NeuroCore.slicedim — Function

slicedim(x)
slicedim!(x, val)

Which dimension slices where acquired at throughout MRI acquisition.

NeuroCore.slice_start — Function

slice_start(x)
slice_start!(x, val)

Which slice corresponds to the first slice acquired during MRI acquisition (i.e. not padded slices).

NeuroCore.slice_end — Function

slice_end(x)
slice_end!(x, val)

Which slice corresponds to the last slice acquired during MRI acquisition (i.e. not padded slices).

NeuroCore.slice_duration — Function

slice_duration(x)
slice_duration!(x, val)

Time to acquire one slice

Spatial Encoding

Properties

NeuroCore.nshots — Function

nshots(x)
nshots!(x, val)

The number of RF excitations needed to reconstruct a slice or volume. Please mind that this is not the same as Echo Train Length which denotes the number of lines of k-space collected after an excitation.

NeuroCore.effective_echo_spacing — Function

effective_echo_spacing(x)
effective_echo_spacing!(x, val)

The "effective" sampling interval, specified in seconds, between lines in the phase-encoding direction, defined based on the size of the reconstructed image in the phase direction. It is frequently, but incorrectly, referred to as "dwell time" (see dwell_time parameter below for actual dwell time). It is required for unwarping distortions using field maps. Note that beyond just in-plane acceleration, a variety of other manipulations to the phase encoding need to be accounted for properly, including partial fourier, phase oversampling, phase resolution, phase field-of-view and interpolation.<sup>2</sup> This parameter is REQUIRED if corresponding fieldmap data is present.

<sup>2</sup>Conveniently, for Siemens’ data, this value is easily obtained as 1/[BWPPPE * ReconMatrixPE], where BWPPPE is the "BandwidthPerPixelPhaseEncode in DICOM tag (0019,1028) and ReconMatrixPE is the size of the actual reconstructed data in the phase direction (which is NOT reflected in a single DICOM tag for all possible aforementioned scan manipulations). See here and here

NeuroCore.parallel_acquisition_technique — Function

parallel_acquisition_technique(x)
parallel_acquisition_technique!(x, val)

The type of parallel imaging used (e.g. GRAPPA, SENSE).

NeuroCore.parallel_reduction_factor_in_plane — Function

parallel_reduction_factor_in_plane(x)
parallel_reduction_factor_in_plane!(x, val)

The parallel imaging (e.g, GRAPPA) factor. Use the denominator of the fraction of k-space encoded for each slice. For example, 2 means half of k-space is encoded.

NeuroCore.partial_fourier — Function

partial_fourier(x)
partial_fourier!(x, val)

The fraction of partial Fourier information collected.

NeuroCore.partial_fourier_direction — Function

partial_fourier_direction(x)
partial_fourier_direction!(x, val)

The direction where only partial Fourier information was collected.

NeuroCore.total_readout_time — Function

total_readout_time(x)
total_readout_time!(x, val)

The total readout time. This is actually the "effective" total readout time , defined as the readout duration, specified in seconds, that would have generated data with the given level of distortion. It is NOT the actual, physical duration of the readout train. If effective_echo_spacing has been properly computed, it is just effective_echo_spacing * (ReconMatrixPE - 1).<sup>3</sup> .

This parameter is

REQUIRED if corresponding "field/distortion" maps acquired with opposing phase encoding directions are present.

<sup>3</sup>We use the "FSL definition", i.e, the time between the center of the first "effective" echo and the center of the last "effective" echo.

Magentization Transfer

NeuroCore.MagnetizationTransferMetadata — Type

MagnetizationTransferMetadata

Metadata structure for information concerning MRI magnetization transfer pulse.

Supported Properties

mt_state: Specifies whether the magnetization transfer pulse is applied. This parameter is REQUIRED by all the anatomical images grouped by MTR, MTS and MPM suffixes.

mt_offset_frequency: The frequency offset of the magnetization transfer pulse with respect to the central H1 Larmor frequency in Hertz (Hz).

mt_pulse_bandwidth: The excitation bandwidth of the magnetization transfer pulse in Hertz (Hz).

mt_npulses: Number of magnetization transfer RF pulses applied before the readout.

mt_pulse_shape: Shape of the magnetization transfer RF pulse waveform. Accepted values:

mt_pulse_duration: Duration of the magnetization transfer RF pulse in seconds.

Properties

NeuroCore.mt_state — Function

mt_state(x) -> Bool
mt_state!(x, val)

Specifies whether the magnetization transfer pulse is applied. This parameter is REQUIRED by all the anatomical images grouped by MTR, MTS and MPM suffixes.

NeuroCore.mt_offset_frequency — Function

mt_offset_frequency(x)
mt_offset_frequency!(x, val)

The frequency offset of the magnetization transfer pulse with respect to the central H1 Larmor frequency in Hertz (Hz).

NeuroCore.mt_pulse_bandwidth — Function

mt_pulse_bandwidth(x)
mt_pulse_bandwidth!(x, val)

The excitation bandwidth of the magnetization transfer pulse in Hertz (Hz).

NeuroCore.mt_npulses — Function

mt_npulses(x)
mt_npulses!(x, val)

Number of magnetization transfer RF pulses applied before the readout.

NeuroCore.mt_pulse_shape — Function

mt_pulse_shape(x)
mt_pulse_shape!(x, val)

Shape of the magnetization transfer RF pulse waveform. Accepted values:

HARD

GAUSSIAN

GAUSSHANN (gaussian pulse with Hanning window)

SINC

SINCHANN (sinc pulse with Hanning window)

SINCGAUSS (sinc pulse with Gaussian window)

FERMI

NeuroCore.mt_pulse_duration — Function

mt_pulse_duration(x)
mt_pulse_duration!(x, val)

Duration of the magnetization transfer RF pulse in seconds.

Spoiling

NeuroCore.SpoilingMetadata — Type

SpoilingMetadata

Metadata structure for information concerning MRI sequence spoiling.

Supported Properties

spoiling_state: Specifies whether the pulse sequence uses any type of spoiling stratey to suppress transverse magnetization remaining after the readout.

spoiling_type: Specifies which spoiling method(s) are used by a spoiled sequence. Accepted values: RF, GRADIENT or COMBINED.

spoiling_gradient_moment: Zeroth moment of the spoiler gradient lobe in militesla times second per meter (mT.s/m).

spoiling_gradient_duration: The duration of the spoiler gradient lobe in seconds. The duration of a trapezoidal lobe is defined as the summation of ramp-up and plateu times.

Properties

NeuroCore.spoiling_state — Function

spoiling_state(x)
spoiling_state!(x, val)

Specifies whether the pulse sequence uses any type of spoiling stratey to suppress transverse magnetization remaining after the readout.

NeuroCore.spoiling_type — Function

spoiling_type(x)
spoiling_type!(x, val)

Specifies which spoiling method(s) are used by a spoiled sequence. Accepted values: RF, GRADIENT or COMBINED.

NeuroCore.spoiling_gradient_moment — Function

spoiling_gradient_moment(x)
spoiling_gradient_moment!(x, val)

Zeroth moment of the spoiler gradient lobe in militesla times second per meter (mT.s/m).

NeuroCore.spoiling_gradient_duration — Function

spoiling_gradient_duration(x)
spoiling_gradient_duration!(x, val)

The duration of the spoiler gradient lobe in seconds. The duration of a trapezoidal lobe is defined as the summation of ramp-up and plateu times.

Time

NeuroCore.echo_time — Function

echo_time(x)
echo_time!(x, val)

The echo time (TE) for the acquisition. This parameter is REQUIRED if corresponding fieldmap data is present or the data comes from a multi echo sequence.

NeuroCore.inversion_time — Function

inversion_time(x)
inversion_time!(x, val)

Returns the inversion time (TI) for the acquisition, specified in seconds. Inversion time is the time after the middle of inverting RF pulse to middle of excitation pulse to detect the amount of longitudinal magnetization.

NeuroCore.slice_timing — Function

slice_timing(x)
slice_timing!(x, val)

The time at which each slice was acquired within each volume (frame) of the acquisition. Slice timing is not slice order – rather, it is a list of times (in JSON format) containing the time (in seconds) of each slice acquisition in relation to the beginning of volume acquisition. The list goes through the slices along the slice axis in the slice encoding dimension (see below). Note that to ensure the proper interpretation of the slice_timing field, it is important to check if the OPTIONAL slice_encoding_direction exists. In particular, if slice_encoding_direction is negative, the entries in slice_timing are defined in reverse order with respect to the slice axis (i.e., the final entry in the slice_timing list is the time of acquisition of slice 0). This parameter is REQUIRED for sparse sequences that do not have the delay_time field set. In addition without this parameter slice time correction will not be possible.

NeuroCore.dwell_time — Function

dwell_time(x)
dwell_time!(x, val)

Actual dwell time (in seconds) of the receiver per point in the readout direction, including any oversampling. For Siemens, this corresponds to DICOM field (0019,1018) (in ns). This value is necessary for the optional readout distortion correction of anatomicals in the HCP Pipelines. It also usefully provides a handle on the readout bandwidth, which isn’t captured in the other metadata tags. Not to be confused with effective_echo_spacing, and the frequent mislabeling of echo spacing (which is spacing in the phase encoding direction) as "dwell time" (which is spacing in the readout direction).

NeuroCore.delay_time — Function

delay_time(x)
delay_time!(x, val)

Returns the user specified time (in seconds) to delay the acquisition of data for the following volume. If the field is not present it is assumed to be set to zero. Corresponds to Siemens CSA header field ldelaytimeInTR. This field is REQUIRED for sparse sequences using the `repetitiontimefield that do not have the slice_timing field set to allowed for accurate calculation of "acquisition time". This field is mutually exclusive withvolume_timing`.

NeuroCore.acquisition_duration — Function

acquisition_duration(x)
acquisition_duration!(x, val)

Duration (in seconds) of volume acquisition. This field is REQUIRED for sequences that are described with the volume_timingfield and that do not have the slice_timing field set to allowed for accurate calculation of "acquisition time". This field is mutually exclusive with repetition_time.

NeuroCore.volume_timing — Function

volume_timing(x)
volume_timing!(x, val)

Returns the time at which each volume was acquired during the acquisition. It is described using a list of times (in JSON format) referring to the onset of each volume in the BOLD series. The list must have the same length as the BOLD series, and the values must be non-negative and monotonically increasing. This field is mutually exclusive with repetitiontime and delaytime. If defined, this requires acquisition time (TA) be defined via either slicetiming or acquisitionduration be defined.

NeuroCore.repetition_time — Function

repetition_time(x)
repetition_time!(x, val)

Returns the time in seconds between the beginning of an acquisition of one volume and the beginning of acquisition of the volume following it (TR). Please note that this definition includes time between scans (when no data has been acquired) in case of sparse acquisition schemes. This value needs to be consistent with the pixdim[4] field (after accounting for units stored in xyztunits field) in the NIfTI header. This field is mutually exclusive with `volumetiming` and is derived from DICOM Tag 0018, 0080 and converted to seconds.